CONSTANTS: adding constants for the scoring system GameStatus: adding a game status to mark a game as finished ActionType: adding a method in the interface Guess: let it implement the interface Hint: let it implement the interface Term: let it implement the interface Player: changed the type of elapsedTime to use in the scoringSytem PlayerGetDTO: changed the type of elapsedTime PlayerPutDTO: changed the type of elapsedTime GameService: deleted some unused methods PlayerService: changes due to the scoringSystem, adjusted the update playermethod to prevent setting ready/notready after game started RoundService: Added setting time functionality and put start and end time into correct order ValidationService: added the updateScores for the players and adjusted the check guess validity Tests: adjusted tests to new behaviour GameStatus: deleted old GameStatus
FS20G17-62
FS10G17-14
FS10G17-13

+++ src/main/java/ch/uzh/ifi/seal/soprafs20/entity/Player.java
@@ -43,7 +43,7 @@ public class Player implements Serializable {
     private String userToken; //is equal to the user's token
 
     @Column(nullable = false)
-    private Double elapsedTime;
+    private long elapsedTime;
 
 
     public Long getId() {
@@ -110,11 +110,11 @@ public class Player implements Serializable {
         this.userToken = userToken;
     }
 
-    public Double getElapsedTime() {
+    public long getElapsedTime() {
         return elapsedTime;
     }
 
-    public void setElapsedTime(Double elapsedTime) {
+    public void setElapsedTime(long elapsedTime) {
         this.elapsedTime = elapsedTime;
     }
 }
